Update time is now reported in MST

+++ run.sh
@@ -10,9 +10,11 @@ LOG="../serverLog.txt"
 PID=".codb_pid.txt"
 
 killProc () {
-	echo "Stopping server process."
-	kill -9 $(cat $PID)
-	rm $PID
+	if [ -f $PID ]; then
+		echo "Stopping server process."
+		kill -9 $(cat $PID)
+		rm $PID
+	fi
 }
 
 runHost () {
@@ -27,9 +29,8 @@ runHost () {
 helpText () {
 	echo "Runs hosting server for the comparative oncology database."
 	echo ""
-	echo "start	Runs host on local server to be proxied by nginx."
+	echo "start	Kills running processes and starts new server"
 	echo "stop	Kills process using pid in $PID."
-	echo "restart	Kills running processes and starts new server."
 	echo "help	Prints help text and exits."
 	echo ""
 }
@@ -38,11 +39,9 @@ if [ $# -eq 0 ]; then
 	helpText
 elif [ $1 = "help" ]; then
 	helpText
-elif [ $1 = "start" ]; then
-	runHost
 elif [ $1 = "stop" ]; then
 	killProc
-elif [ $1 = "restart" ]; then
+elif [ $1 = "start" ]; then
 	killProc
 	runHost
 fi
